+++ src/ui/dock.h
@@ -31,6 +31,7 @@ private:
 	QLabel *detector_;
 	QLabel *near_ = nullptr;
 	QLabel *update_ = nullptr;
+	QLabel *locked_ = nullptr;
 	QCheckBox *closest_ = nullptr;
 	QPushButton *dual_ = nullptr;
 	QComboBox *active_;